+++ src/test/java/com/github/ddth/recipes/test/checkpoint/TestFunctionalJdbcCheckpointDao.java
@@ -1,16 +1,14 @@
 package com.github.ddth.recipes.test.checkpoint;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-
-import javax.sql.DataSource;
-
-import org.springframework.jdbc.datasource.SingleConnectionDataSource;
-
 import com.github.ddth.dao.jdbc.IJdbcHelper;
 import com.github.ddth.dao.jdbc.impl.DdthJdbcHelper;
 import com.github.ddth.recipes.checkpoint.ICheckpointDao;
 import com.github.ddth.recipes.checkpoint.jdbc.JdbcCheckpointDao;
+import org.springframework.jdbc.datasource.SingleConnectionDataSource;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.DriverManager;
 
 public class TestFunctionalJdbcCheckpointDao extends BaseFunctionalCheckpointTest {
     static class MyCheckpointDao extends JdbcCheckpointDao {
@@ -25,11 +23,10 @@ public class TestFunctionalJdbcCheckpointDao extends BaseFunctionalCheckpointTes
     @SuppressWarnings("resource")
     @Override
     protected ICheckpointDao initCheckpointDao() throws Exception {
-        // if (System.getProperty("enableTestsMySql") == null
-        // && System.getProperty("enableTestsMySQL") == null
-        // && System.getProperty("enableTestsMYSQL") == null) {
-        // return null;
-        // }
+        if (System.getProperty("enableTestsMySql") == null && System.getProperty("enableTestsMySQL") == null
+                && System.getProperty("enableTestsMYSQL") == null) {
+            return null;
+        }
         String jdbcUrl = System.getProperty("jdbc.url", "jdbc:mysql://localhost:3306/test");
         String jdbcUser = System.getProperty("jdbc.user", "test");
         String jdbcPassword = System.getProperty("jdbc.password", "test");